requirements.txt
streamlit
google-generativeai
python-dotenv
pandas
fuzzywuzzy
python-Levenshtein
